Aplicação dividir funcional

+++ Documents/Estágio/PrimeiroProjetoReact/backend/linaPDF/methods/merge.py
@@ -1,5 +1,6 @@
 import os
 from PyPDF2 import PdfFileMerger
+from django.core.files.storage import FileSystemStorage
 
 def merge(arquivos):
 
@@ -10,8 +11,9 @@ def merge(arquivos):
         
     merger.merge(position= index+1, fileobj=arquivos[index+1])
 
-    # Write to an output PDF document
-    os.remove("document-output.pdf")
+    fs = FileSystemStorage()
+    if fs.exists("document-output.pdf"):
+        os.remove("document-output.pdf")
     output = open("document-output.pdf", "wb")
     merger.write(output)
     output.close()
